requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
spotipy==2.4.4
urllib3==1.25.7
